requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2023.12.2
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.3
requests==2.31.0
scipy==1.11.4
sqlparse==0.4.4
sympy==1.12
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
